revert a81eb6c and instead just check the length
We need to serialize and deserialize the result string, so it's not
safe to just truncate it. We could be really specific and compute the
max message length, etc., but for now we will simply insist that the
serialized result is below the limit and ask the user to figure out how
to make that happen by shortening their message, etc.