Fix bug with page token
ListActivityTypesRequest gets mutated when we list activity types;
in particular, it may be set with a nextPageToken.
This request object then gets re-used on the next list activity types request,
which can result in an inconsistent page token error.
We should use a new request object instead of re-using the same one.